pyproject.toml
[project]
name = "emap"
version = "0.1.0"
description = "Evolution under Multi-Agent Pressure: Resource-Constrained Evolution of Multi-Agent Programming Architectures"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Anonymous", email = "anonymous@example.com"}
]
keywords = [
    "evolutionary-algorithms",
    "multi-agent-systems",
    "llm",
    "code-generation",
    "resource-constraints"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "deap>=1.4.0",
    
    # Data & Analysis
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    
    # Async & Parallelization
    "aiohttp>=3.8.0",
    
    # Type checking
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "black>=23.9.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]
tracking = [
    "mlflow>=2.7.0",
    "wandb>=0.15.0",
]

[project.scripts]
emap-evolve = "emap.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/emap"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
